alterações em classes para inclusao do get de rotas por api

+++ app/src/main/java/br/com/seller66/model/ItemPedido.java
@@ -4,7 +4,7 @@ import java.io.Serializable;
 
 public class ItemPedido implements Serializable {
     private Produto produto;
-    private float quantidade;
+    private int quantidade;
 
     public Produto getProduto() {
         return produto;
@@ -14,11 +14,11 @@ public class ItemPedido implements Serializable {
         this.produto = produto;
     }
 
-    public float getQuantidade() {
+    public int getQuantidade() {
         return quantidade;
     }
 
-    public void setQuantidade(float quantidade) {
+    public void setQuantidade(int quantidade) {
         this.quantidade = quantidade;
     }
 }
